Reform main function to adapt custom resource registry
